Extract dense landmark reduction into its own workspace module
